[bazel] update `build_otbn.py` to import otbn_as module
The `build_otbn.py` was invoking the OTBN assembler Python script via
the shell, instead of importing the module directly. This does not
enable bazel make use of the hermetic python toolchain and packages.
This updates the `build_otbn.py` to invoke the OTBN assembler Python
script via a direct function call from a module import.

Signed-off-by: Timothy Trippel <ttrippel@google.com>